# public/openapi.yaml
openapi: 3.0.0
info:
  title: Vitto API
  version: 1.0.0
  description: API documentation for Vitto Next.js backend endpoints
servers:
  - url: /api
paths:
  /auth/send-otp:
    post:
      summary: Send OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OTP sent successfully
        '400':
          description: Invalid email address
        '500':
          description: Failed to send OTP
  /auth/verify-otp:
    post:
      summary: Verify OTP and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
      responses:
        '200':
          description: OTP verified successfully
        '400':
          description: Invalid OTP or expired
        '500':
          description: Failed to verify OTP
  /leads:
    post:
      summary: Create a new lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                institutionName:
                  type: string
                institutionType:
                  type: string
                city:
                  type: string
                phoneNumber:
                  type: string
                loanBookSize:
                  type: string
      responses:
        '200':
          description: Account created successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Failed to create account
    get:
      summary: Get lead by ID
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: Lead ID
      responses:
        '200':
          description: Lead found
        '400':
          description: Missing lead ID
        '401':
          description: Unauthorized
        '500':
          description: Failed to fetch lead
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT